fix(pas): avoid `<sym>: .BLOCK 0`  in macro expansions
Macro invocations as supposed to work __as if__ normal mnemonics.
This includes allowing symbol declarations.

Macros themselves may define symbols on their first line.
Additionally, the first line of a macro may be a comment or a newline, not just addressable code.
Both statements tend to be true for library routines like `@MALLOC`.
Both prevent naively putting a symbol on the first IR line of the macro via naive textual substitution.
We are unwilling to place such severe restrictions on macro authors because it diminished their utility.

This commit modifies the assembler to move the symbol from the macro invocation to the first addressable line if possible.
Only when this fails, e.g. `@MALLOC`, do we resort to using a `<sym>: .BLOCK 0` construct at the top of the macro.
We decided that we don't want the "extra" symbol to disappear in cases like `@MALLOC` and have kept the BLOCK construct.

Closes #602.