feat: add "save _ as" menu items
Which items are present depends on the loaded project.
Fixed a bug where menu items would refer to an outdated project from lack of Qt.binding nonsense.
Still need to actually _do_ save as, but this is progress.

No progress on WASM yet, since that uses a fundamentally different file persistence mechanism.