feat: prompt to save on application quit
Use of recursion-via-chained-signals is because dialogs reenter the event loop and are thus async.
Unify onCloseAllProjects && onQuit to use a common recursion method.

This recursion method will not handle cancel nor keeping some tabs open.
I think I want a dialog that looks more like Qt's, but this is a start.