refactor!: Use IR[8:23] in operand specifier decoding
Further progress on #878.
OldPC4 math has been eliminated in favor of using operand specifier directly.

The OS has two "kinds" of system calls: those acting as loads and those acting as stores.
The relationship between the decoder operand specifier for loads and stores is as follows:
`Operand_load = Mem[Operand_store]`
Given this relationship, decoding to a store operand is more universal and is what we already compute in the old OS.

Note that immediate addressing cannot exist for stores, because `Operand_load = OS`, which does not have a memory component.
Using this info, I was able to combine I and D decoding paths.

BREAKING CHANGE: Changes OS memory map