feat(debug): add runtime type registry
This thing is designed so that we don't have to "carry" runtime tpye into in all of our value classes.
For non-primitive types, this data can be massive due to multiple pointers.
This means EVERY `Type` will be massive.
So I'm going to store the "real" type info in my RTTI registry, and Values will keep a pointer to them.