refactor: replace TypedBits with value variants
RTTI is not plumbed throughout, and type evaluations are not parsed.
Pretty sure this will fail tests, but I want to see if it compiles!!

Also have some fixes for integer rank promotion.
Seems it never really worked, but passed unit tests.
I've refactored some of the logic to avoid nested conditionals.