feat(debug): serialize stack tracing operations
Still need to deserialize.
For now, lots of output to cerr to validate different programs.

MemoryOp uses a string + BoxedType now.
I don't have an API for DirectHandle->quint16, and it would be difficult to write one.

These opcodes are meant to be easy to execute.
Moving the actual types into the op reduce the # of things that need to be passed to the stack tracer.