fix(help): prevent broken QRC references in Help
Executables built via GitHub can't access any of the images in the qrc, but my locals builds can?

I tracked it down to the start of my URLs being `qrc:/` and not some relative path.
No clue why this fixes my problem, but it fits with my mental model of QRC being semi-janky.