fix: keep OS AST alive as long as user program
Was hitting a bug where OS symtab was reclaimed before use in formatting user program.
It only occurred in the `tex` subcommand due to the extended lifetime of the user program over the assembler.
0420 has been updated in a way which triggers this behavior (usage of OS symbols).

All output from the assembler utils now extend the lifetime of the OS using a "KeepAlive" property.
I've added logging to the offending path for future debugging efforts.

It's a hack, but we have a deadline.