fix: partially enable undo/redo in assembly editor
I'm having problems where all of my events are being coalesced.
I don't have the time to figure it out today.

I now insert some uncoalescable undo actions on space, tab, and newline.
The downside is that these actions require two undo keypresses to skip over.

I expect the problem is that I open a group and do not close it.
But I have not been able to trace that down yet.
Closes #922 for now.