fix(welcome): also copy figure inputs from "favorites"
Was not copying inputs before.
Needed to plumb setCharIn from the recent files all the way to main.qml

In the help system, Figure2.qml's signals are delegated to the HelpRoot.
In the case of adding projects, HelpRoot "unpacks" the one add project call into multiple signals.
Since we aren't part of HelpRoot, so we need to match the signature expected in main.qml