feat(pas): add start of linear IR representation for Pep/10 ASM
Next step is to write a parser which emits this IR.
Then, I can write some code which "eats" this IR and converts it to my old tree style.

As the new assembler gets better, I can do that conversion later and later.
Eventually, no conversion will be necessary, and I can remove the conversion code + the old IR.