fix(bi): replace end index with length
`mid` takes an (offset, length) pair, not (offset, offset).
I've fixed this by computing the length between the two offsets
Effectively, I was passing a value much larger than length when I had a non-0 start.

I also have an off-by-1 for negative indices.
I did not record the end of the string as a final lineStarts position.
When I fixed the first bug, this second bug started to manifest due to mid not asserting boundaries.
I've stripped line from the loop variable because it was unused.