refactor: change default layout of of messages, IO panes
This removes the need to tab back and forth to view message contents.
The new order needs to be stable despite the fact the batch output does not exist in the "Editor" mode.

I'm running into some issues with the actual pixel math of getting each pane to be exactly the correct size.
I will need to tweak this math in the future.
I already plan on revisiting my layouts (#855), and these relationships can be re-evaluated at that time.

Progress on #965.