refactor(proj): retain assembler messages when running Asmb projects
Run is really three operations wearing a trench coat: assemble, load, execute.
Each of these ops wants to clear messages if it the top level interaction with the user.

Because I lacked this detection, running "RET" at ASMB5 would cause the messages tab to flash.
When you switched to it, there would be no messages!
Assembly would post a "success", which would be promptly deleted by execute, leading to a confusing UX.

onAssemble is the only offender at the moment.
So, I've created a wrapper which is exposed to the UI which does the clearMessages, and the guts which does the assembly.
The guts must not call clearMessages.

I've also removed what I think are spurious calls to clearMessages.

Should close #969.