feat(debug): propagate current SP to StackTracer
Computer Systems 6th edition allows (read as: encourages, requires) self modifying programs.

This can take three forms relevant to stack tracing:
 (a) Replace a stack modifying instruction with a different instruction
 (b) Replace a non-stack instruction with a stack instruction
 (c) Modify the number of bytes de/allocated via operand specifier replacement

Cases (a) and (b) are covered by the simulator posting events to the stack tracing SW on stack modifying instructions.
The event could be compared against the trace commands.
If the trace command does not make sense for the instruction (e.g., deactivate a frame on an alloc), we can halt tracing / enter an invalid state.
Alternatively, if there is no event, we can enter an invalid state.

These methods to not cover (c), since the tracer is unaware of the number of bytes actually allocated.
So, the simulator needs to notify the debugger of the new SP after each instruction.
The tracer can compare the number of bytes actually de/allocated to the trace command to catch errors.

These defenses should prevent self-modifying programs from segfaulting Pepp.

This has the unfortunate side effect of being very non-physical in terms of implementation.
The previous PC and future SP are signaled from the CPU to the debugger.
That is, we are observing partially committed CPU state, which now source-level-debugger should be given.