feat(debug): add new expression node to optimize `*(T*)<Expr>`
Usually, a cast - deref would generate ~4 AST nodes, 3 of which would be volatile.
An inner expression would be created for the address "Expr"
Then, we would wrap that in a MemoryRead
We would then need a DirectCast (cast the pointer from MemoryRead to the desired type's pointer).
Then we would need a deref which would strip the outer * and get us to the target type.
Those last 3 nodes would all be volatiles that need to be updated.

I've created an optimized node which combines these operations.
Its primary purpose is to support stack cell evaluation, but it may be useful elsewhere.
A future edit may optimize the pattern from my commit message into this new node.