fix(pas): initialize DotEquate IR with symbol declaration
Propagate the (optional) symbol through the parser functions.
This way we can ensure that there are no partially constructed IR nodes.

Otherwise I would need the symbol to be optional on .EQUATE and I don't want that added complexity.
It would need to be optional because attribute would return a pointer to an (empty) symbol.
The rest of the parser would need to guard against this "uninitialized" symbol.