refactor(pas): create simple, inlline visitor pattern for IR types
All IR lines now have a TYPE, which must be unique for each class.
It is accessed via the virtual method type(), following a similar pattern in the expr parser.
You can switch(type()) and then static_<pointer_>cast to the underlying type.

This is less type safe than a bunch of dynamic pointer casts with check.
However, it has the benefit of being O(1) performance in the number of checks vs O(n).

This is not a true visitor, because the behavior is not encapsulated.
As I add future behaviors (e.g., object code output) this pattern may need to evolve.

I've also combined input/output/import/export/scall.
All provide some assembly-time hint to the code generator.
None generate object code, and all have the same argument types.