feat(pas): add `format_source` equivalent for new assembler
Operates on tokens rather than IR.
The new assembler will emit a flat IR where the macros are inlined via magic from the lexer.
This has a benefit of the formatting working on a large variety of programs than `Assemble & Format` (e.g., programs with syntax & semantic errors).

We can snoop on the top level lexer to get just the top level tokens.
Then we can format source code while completely ignoring macros.

It will be a bit of a pain, but I think the listing can be made to take advantage of this source formatting.