fix(pas): correct symbol section indices when skipping sections
When we pre-computed section indices we did not know if some sections would be skipped.

This could be solved by adding another loop at the end of assign_addresses which updates lines' symbol's section indices.
Or we could record which sections need a fixup and apply that fix when writing out the symbol table, which is what I have opted for.

We already have to process each symbol, so this approach does not create any additional loops over the IR.