refactor(pas): replace StringPool with hash map
Using a hash map now, whose elements are stable across insertions even if iterators are invalidated.
That is arr.find("key") can be invalidated, while &*arr.find("key") will not be invalidated.
I had looked into tries, but they often invalidate items on insert, which would necessitate pre-compute the set of possible identifiers.
I don't want to iterate over the input twice, so hash maps (e.g., std::unordered_set)  it is.

The StringPool basically tried to cram as many strings into a page / line as possible, creating substring views where possible.
Originally designed to reduce memory usage, it has a relatively high performance cost.
That is, it may need to iterate over many strings on insert.
The possible memory savings are small (orders of kilobytes) and not worth it.
The StringPool also did not end up being useful for "compressing" elf string sections because of the fixed-size pages.

I generally think it's going to be a combinatorial optimization problem to optimally shared elf strings.
My gut says this is NP-hard, and this data structure is probably the wrong choice.