refactor: vendor `libriscv` into simulator directory
I plan on modifying my simulator API to have a tracing mode and a nontracing mode.
I want to adapt both my Pep/N and the new libriscv simulators to support both APIs.

This either requires inlining libriscv, or extracting the simulator API; I have chosen the former.
Tests are not yet ported over, but I plan on pre-compiling the `C` code to compact ELF files to reduce the number of dependencies in the CI environment.

When this vendored code is comparable to functionality to my libriscv fork, I will archive that repo.